rename file_asset to vector_file_asset
