+++ cv-assets/cv_assets/vectors/mndnr_culvert_inventory.py
@@ -4,7 +4,7 @@ from string import Template
 from dagster import asset
 
 from cv_assets.config import get_settings
-from cv_assets.resources.file_asset import VectorFileAsset
+from cv_assets.resources.vector_file_asset import VectorFileAsset
 
 settings = get_settings()
 TARGET_EPSG = settings.target_epsg
